requirements.txt
torch==2.0.1
torchvision==0.15.2
timm==0.9.12
gradio==4.7.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
matplotlib==3.8.0
tqdm==4.66.1
wget
pandas